Added samples for Copilot and Dynamics integration and launch canvas app from sitemap
These samples will show you how to launch a side pane with a Copilot Studio agent and pass context. This will allow the agent to analyze and present information to users. The other sample shows how to launch a canvas app from the site map. This allows parameters to be passed in, specifically session id and other correlation id to connect app insights telemetry.